- (feat)Implement SurrealDB connexion and clerk integration with our db we are now connected to our surreal cloud database. We can auth with the token provided by clerk, and if the user clerk send us is not in our db yet, we will create it with basic information provided by our token. We had to create special field in clerk to get those information so that we can save them directly in our db and never use clerk again to fetch data about our user.
- Implement a proper connecttion to our db at the root and write the first
query
We have create a new Provider, works more like a context, the goal is to
create the connecttion, to the db so that our app can use the client and
query. This connection happens on_mount for this
component/provider/context. I neede to figure out what I want to do with
this.
We have returned the first list of user related to our organisation and
the permissions of our database works, we can only see and return client
related to our organisation without having to specify a filter, this
allow us to query directly from the client.